Room mirroring for test room complete
Notable problems:
Objects that have placement dependent on their width will not be placed in the same place when the room is mirrored as they are in rooms that aren't mirrored. If we make it so that the center of all objects is independent of mirroring, this will be fixed.